Remove EcdsaPrivateKey from scrypto module
